RD-2005-10-13-0: Bugfix:    fixed broken builds                  Review:    MD

+++ source/libs/sgeobj/test_sge_schedd_conf.c
@@ -98,7 +98,7 @@ test(conf_settings_t *setting, schedd_conf_t *test, int test_counter)
 
       switch (test->type) {
          case lStringT :
-               lSetString(schedd_conf, conf_tests->test_attribute, setting->test_value);
+               lSetString(schedd_conf, (int)conf_tests->test_attribute, setting->test_value);
             break;
          case lDoubleT :
             break;
